# Conflicts:
#	Triangulation_3/examples/Triangulation_3/CMakeLists.txt
#	Triangulation_3/include/CGAL/Triangulation_3.h
#	Triangulation_3/include/CGAL/Triangulation_utils_3.h

+++ Arrangement_on_surface_2/include/CGAL/Arr_rat_arc/Cache.h
@@ -14,12 +14,16 @@
 //
 // $URL$
 // $Id$
+// SPDX-License-Identifier: GPL-3.0+
 //
 // Author(s)     : Oren Salzman <orenzalz@post.tau.ac.il >
 //                 Michael Hemmer <Michael.Hemmer@sophia.inria.fr>
 
 #ifndef CGAL_RATIONAL_ARC_CACHE
 #define CGAL_RATIONAL_ARC_CACHE
+
+#include <CGAL/license/Arrangement_on_surface_2.h>
+
 
 #include <CGAL/Arr_rat_arc/Base_rational_arc_ds_1.h>
 #include <CGAL/Arr_rat_arc/Rational_function.h>
@@ -277,10 +281,9 @@ private:
     }
 
     //re-set watermark
-    _rat_func_map_watermark = (std::max)(
-        2*_rat_func_map.size(),
-        typename Rational_function_map::size_type(128));
-    
+    _rat_func_map_watermark
+      = static_cast<unsigned int>((std::max)(2* _rat_func_map.size(),
+                                             typename Rational_function_map::size_type(128)));
     return;
   }
   void rat_pair_map_clean_up() const 
@@ -307,8 +310,8 @@ private:
 
     //re-set watermark
     _rat_pair_map_watermark = 
-      (std::max)(2*_rat_pair_map.size(),
-          typename Rational_function_canonicalized_pair_map::size_type(128));
+      static_cast<unsigned int>((std::max)(2* _rat_pair_map.size(),
+                                           typename Rational_function_canonicalized_pair_map::size_type(128)));
   }
 
 private:
