+++ Straight_skeleton_2/include/CGAL/Straight_skeleton_2/Straight_skeleton_builder_2_impl.h
@@ -13,11 +13,15 @@
 //
 // $URL$
 // $Id$
+// SPDX-License-Identifier: GPL-3.0+
 //
 // Author(s)     : Fernando Cacciola <fernando_cacciola@ciudad.com.ar>
 //
 #ifndef CGAL_STRAIGHT_SKELETON_BUILDER_2_IMPL_H
 #define CGAL_STRAIGHT_SKELETON_BUILDER_2_IMPL_H 1
+
+#include <CGAL/license/Straight_skeleton_2.h>
+
 
 #include <boost/bind.hpp>
 #include <boost/utility.hpp>
@@ -540,7 +544,7 @@ void Straight_skeleton_builder_2<Gt,Ss,V>::CreateContourBisectors()
 template<class Gt, class Ss, class V>
 void Straight_skeleton_builder_2<Gt,Ss,V>::InitPhase()
 {
-  mVisitor.on_initialization_started(mSSkel->size_of_vertices());
+  mVisitor.on_initialization_started(static_cast<int>(mSSkel->size_of_vertices()));
   CreateContourBisectors();
   CreateInitialEvents();
   mVisitor.on_initialization_finished();
